refactor(op): SerializeGraphs — the framework owns the graph-stream rendering
Single-codec Requirement 1 (ruled 2026-08-04): pkg/op must own both
directions of graph-document handling, including every serialization
aspect. op.SerializeGraphs(w io.Writer, graphs []*op.Graph) lands
beside op.LoadGraph as the write-side complement — two-space indent,
multi-document YAML framing, encoder-close error folding — and the
three byte-identical emitGraphs copies in writ deploy, upgrade, and
decommission are deleted; each dry-run path reduces to one call with
the destination writer as the only caller choice. Plan ledger:
docs/plans/single-codec.md (new); the phase-1 open ruling in
docs/plans/complexity-refactors.md is marked resolved.
