+++ cmd/writ/writ/deploy/deploy.go
@@ -20,8 +20,6 @@ import (
 	"sort"
 	"strings"
 
-	"gopkg.in/yaml.v3"
-
 	"github.com/NobleFactor/devlore-cli/cmd/lore/lore"
 	"github.com/NobleFactor/devlore-cli/cmd/writ/writ/readback"
 	"github.com/NobleFactor/devlore-cli/cmd/writ/writ/segment"
@@ -118,7 +116,7 @@ func Execute(ctx context.Context, cfg *Config) (err error) {
 	}
 
 	if cfg.DryRun {
-		return emitGraphs(build.Graphs)
+		return op.SerializeGraphs(os.Stdout, build.Graphs)
 	}
 
 	sortGraphsByScope(build.Graphs)
@@ -129,32 +127,6 @@ func Execute(ctx context.Context, cfg *Config) (err error) {
 	}
 
 	return runAll(ctx, cfg, build.Graphs, runPolicy)
-}
-
-// emitGraphs serializes the graphs to stdout as one YAML stream — the dry-run rendering.
-//
-// Parameters:
-//   - `graphs`: the assembled graphs, in run order.
-//
-// Returns:
-//   - `err`: a serialization or encoder-close failure.
-func emitGraphs(graphs []*op.Graph) (err error) {
-
-	encoder := yaml.NewEncoder(os.Stdout)
-	defer func() {
-		if closeErr := encoder.Close(); closeErr != nil && err == nil {
-			err = closeErr
-		}
-	}()
-	encoder.SetIndent(2)
-
-	for _, graph := range graphs {
-		if err := graph.Serialize(encoder); err != nil {
-			return err
-		}
-	}
-
-	return nil
 }
 
 // runAll executes every graph under the run policy, collecting per-scope failures.
